requirements.txt
# Core ML/DL libraries
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.20.0

# RL and optimization
gymnasium>=0.28.0
stable-baselines3>=2.0.0
optuna>=3.0.0

# Data processing and utilities
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0
tqdm>=4.65.0

# Configuration and logging
pyyaml>=6.0
hydra-core>=1.3.0
python-dotenv>=1.0.0

# Experiment tracking and visualization
wandb>=0.15.0
tensorboard>=2.13.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilities
requests>=2.31.0
tiktoken>=0.4.0
sentencepiece>=0.1.99
protobuf>=4.23.0

# Development and testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# Optional: For advanced features
# bitsandbytes>=0.41.0  # For 8-bit quantization
# flash-attn>=2.0.0     # For flash attention (requires specific CUDA)
# xformers>=0.0.20      # For memory efficient attention
